requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
openai>=1.99.0
pydantic>=2.11.0
langgraph>=0.3.0
python-multipart>=0.0.9
pypdf>=5.1.0
python-docx>=1.1.2
